PartnerFinder finds closest partners
The partnerfinder uses filter methods to find matching partners. The
callback functions for the Array.filter() methods must return a boolean,
so they check for an office of a partner that's in range and return true
if it is. Because the check is two layers deep in an object, passing out
the 'true' is based on the existence of the inner filter finding an
office within distance, rather than an empty array.